Move mocha JS injecting and phantomjs specific workarounds for mocha to phantomjs platform object. (for screenshot testing lib)
